moving print to config
